Exports are now concatenated monthly exports
As of about 2018-05-26 the database stopped allowing exports of greater than
3000 records. Unfortunately, I didn't catch that change until today. To
work around this limitation, snapshot.py simply requests an export for
each month across a range of years, and concatenates the results into
export.csv. It's not pretty the export data was always going to need
some clean up.